pyproject.toml
[tool.poetry]
name = "pysqlitedb"
version = "0.1.0"
description = "Simple wrapper for sqlite3 db"
authors = ["Venky Iyer <indigoviolet@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
attrs = "^20.3.0"
pytz = "^2020.4"

[tool.poetry.dev-dependencies]
mypy = "^0.790"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
